backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
httpx

# LLM providers (install the ones you need)
anthropic
openai
google-generativeai

# Testing
pytest
pytest-asyncio
httpx
